requirements.txt
python-dotenv>=1.0,<2.0
python-telegram-bot>=20,<22
fabric>=3.2.2
